test: split CocoSketch and Elastic into their own family suite and wire them through the kit
`e2e_experimental.rs` was held together by a cargo feature rather than by
subject. CocoSketch and the Elastic sketch are a family: both keep a flow key
beside every counter and answer heavy-hitter questions off it. They move to
`tests/e2e_keyed_buckets.rs`, named for the same thing `e2e_octo.rs`'s
`keyed_buckets` module is named for -- that file covers the Octo variants of
these two, this one the single-threaded sketches.

Neither was wired through the conformance kit at all. Both now run
`frequency_battery` and `merge_equivalence_battery` behind thin adapters:

- Coco two-sided, at its documented 1024 x 4 default. The API doc calls it
  unbiased and promises no floor, so `one_sided` stays false. The spec is
  Count Sketch's two-sided reference spec from `conformance_kit.rs` unchanged;
  Coco elects from an unseeded RNG, and over 40 runs the widest dense-key
  deviation used 0.7 of the 25.0 absolute floor.
- Elastic one-sided, at 256 heavy buckets over the default 3 x 4096 light
  layer -- §3.1.2 puts a quarter of the buckets in contest at that width, so
  the losers actually read through the light layer. The excess above truth is
  that layer's Count-Min error, so the tolerance is the additive eps * N with
  eps = e / cols and nothing relative. Measured worst excess is 8.

`turnstile_battery` fits neither: no decrement path on either sketch.

The seven moved tests keep their assertions and tolerances. Test totals go
855 -> 857 all-features and 801 -> 803 with `--features experimental`, the
two new battery tests; 686 unchanged with no features.
